+++ tide_app.py
@@ -8,10 +8,10 @@ from fastapi.openapi.utils import get_openapi
 from fastapi.responses import JSONResponse, ORJSONResponse
 from fastapi.encoders import jsonable_encoder
 from typing import Optional #List
-from pydantic import Field #BaseModel,
+from pydantic import Field #validator, BaseModel
 import requests
 import json
-from datetime import date, datetime, timedelta
+from datetime import datetime, timedelta
 from src.model_utils import get_tide_time, get_tide_series, get_tide_map
 import src.config as config
 from dask.distributed import Client
@@ -36,7 +36,7 @@ def generate_custom_openapi():
     return app.openapi_schema
 
 
-app = FastAPI(root_path="/tide", docs_url=None, default_response_class=ORJSONResponse)
+app = FastAPI(root_path="/api/tide", docs_url=None, default_response_class=ORJSONResponse)
 
 
 @app.get("/api/swagger/tide/openapi.json", include_in_schema=False)
@@ -53,7 +53,6 @@ async def custom_swagger_ui_html():
 
 
 ### Global variables: config.py ###
-
 
 def to_global_lonlat(lon: float, lat: float) -> tuple:
     mlon = 180 if lon > 180 else (-180 if lon < -180 else lon)
@@ -151,9 +150,9 @@ async def get_tide(
         None, description="Maximum longitude, range: [-180, 180]"),
     lat1: Optional[float] = Query(
         None, description="Maximum latitude, range: [-90, 90]"),
-    start: Optional[date] = Query(
+    start: Optional[str] = Query(
         None, description="Start datetime (UTC) of tide data to query. If none, current datetime is default"),
-    end: Optional[date] = Query(
+    end: Optional[str] = Query(
         None, description="End datetime (UTC) of tide data to query"),
     sample: Optional[int] = Query(
         5, description="Re-sampling every N points(default 5)"),
@@ -223,6 +222,8 @@ async def get_tide(
 
     tide_time, dtime = get_tide_time(start_date, end_date)
     output_mode = 'time'
+    if mode is None:
+        mode = 'list'
 
     try:
         orig_lon0, orig_lon1 = lon0, lon1
@@ -245,8 +246,8 @@ async def get_tide(
             if findNear:
                 dsub = config.dz.sel(lon=lon0, lat=lat0, method="nearest", tolerance=tol).sel(constituents=cons)
             else:
-                dsub = config.dz.sel(lon=slice(lon0, lon0+1.0*config.gridSz),
-                                     lat=slice(lat0, lat0+1.0*config.gridSz),
+                dsub = config.dz.sel(lon=slice(lon0-0.5*config.gridSz, lon0+0.5*config.gridSz),
+                                     lat=slice(lat0-0.5*config.gridSz, lat0+0.5*config.gridSz),
                                      constituents=cons)
             tide = {}
             for var in variables:
@@ -263,19 +264,15 @@ async def get_tide(
                 tide[var] = ts
         else:
             # Bounding box
-            offset_lat = 0.0
-            offset_lon = 0.0
-            if lat1 == lat0 or abs(lat1 - lat0) < config.gridSz:
-                offset_lat = 1.0 #1.0 means 1.0 * config.gridSz, not in degree
-                #because the to_global_lonlat() not snap to grid, so must +1 gridSz to ensure have at least one point
-                #print("lat0, lat1 equal: ", lat0, lat1)
-
+            # offset_lat = 0.0
+            # offset_lon = 0.0
+            #if lat1 == lat0 or abs(lat1 - lat0) < config.gridSz:
+            #   offset_lat = 0.5 #0.5 means 0.5 * config.gridSz, not in degree
+                #because the to_global_lonlat() not snap to grid, so must expand to ensure have at least one point
+            #if lon1 == lon0 or abs(lon1 - lon0) < config.gridSz:
+            #   offset_lon = 0.5
             if lat1 < lat0:
                 lat0, lat1 = lat1, lat0
-
-            if lon1 == lon0 or abs(lon1 - lon0) < config.gridSz:
-                offset_lon = 1.0
-
             if lon1 < lon0:
                 lon0, lon1 = lon1, lon0
 
@@ -302,14 +299,18 @@ async def get_tide(
                 #    lon0, lon1 = lon1, lon0
                 #    orig_lon0, orig_lon1 = orig_lon1, orig_lon0
                 subset1 = config.dz.sel(
-                    lon=slice(lon0, 360), lat=slice(lat0, lat1+offset_lat*config.gridSz), constituents=cons)
+                    lon=slice(lon0-0.5*config.gridSz, 360),
+                    lat=slice(lat0-0.5*config.gridSz, lat1+0.5*config.gridSz),
+                    constituents=cons)
                 subset2 = config.dz.sel(
-                    lon=slice(0, lon1+1.0*config.gridSz), lat=slice(lat0, lat1+offset_lat*config.gridSz), constituents=cons)
+                    lon=slice(0, lon1+0.5*config.gridSz),
+                    lat=slice(lat0-0.5*config.gridSz, lat1+0.5*config.gridSz),
+                    constituents=cons)
                 ds1 = xr.concat([subset1, subset2], dim='lon')
             else:
                 # Requested area doesn't cross the zero meridian
-                ds1 = config.dz.sel(lon=slice(lon0, lon1+offset_lon*config.gridSz),
-                                    lat=slice(lat0, lat1+offset_lat*config.gridSz),
+                ds1 = config.dz.sel(lon=slice(lon0-0.5*config.gridSz, lon1+0.5*config.gridSz),
+                                    lat=slice(lat0-0.5*config.gridSz, lat1+0.5*config.gridSz),
                                     constituents=cons)
 
             dsub = ds1.isel(lon=slice(None, None, sample), lat=slice(None, None, sample))
@@ -400,17 +401,22 @@ def const_to_output(data_dict): #, data_var='amp'):
 
     return df
 """
-def data_to_wide(df):
-    # Convert to wide format (use methods discussed earlier)
-    # Note wide format conversion cannot allow NA rows
+def data_to_wide(df, mode):
     # df = df.dropna().reset_index(drop=True)
-    # print("Before to wide: ", df)
     # print(df['longitude'].apply(type).unique())
-    wide_format = df.pivot_table(index=['longitude', 'latitude', 'type'], 
+    if 'onlyOnePt' in mode:
+        df['value'] = df['value'].apply(lambda x: x[0] if isinstance(x, list) and len(x) > 0 else x)
+
+    wide_format = df.pivot_table(index=['longitude', 'latitude', 'grid_lon', 'grid_lat', 'type'], 
                                  columns=['constituents', 'variable'], 
                                  values='value').reset_index()
-    col_names = ['longitude', 'latitude', 'type'] + \
-                [f"{c[0].upper()}_{c[1]}"for c in wide_format.columns[3:]]
+    #if 'uppercase' in mode.lower():
+    #    col_names = ['longitude', 'latitude', 'type'] + \
+    #                [f"{c[0].upper()}_{c[1]}"for c in wide_format.columns[3:]]
+    #else:
+    col_names = ['longitude', 'latitude', 'grid_lon', 'grid_lat', 'type'] + \
+                [f"{c[0]}_{c[1]}"for c in wide_format.columns[5:]]
+        
     wide_format.columns = col_names
     return wide_format
 
@@ -440,6 +446,65 @@ def const_to_output(data_dict, mode):
         return data_to_wide(df)
 
     return df
+
+
+def const_to_output_vec(data_dict, mode):
+    data_list = []
+    lon_values = data_dict['longitude']
+    lat_values = data_dict['latitude']
+
+    # Extract keys that are not 'longitude' or 'latitude'
+    other_keys = [key for key in data_dict.keys() if key not in ['longitude', 'latitude', 'grid_lon', 'grid_lat']]
+
+    for idx, (longitude, latitude) in enumerate(zip(lon_values, lat_values)):
+        for key in other_keys:
+            data_type, constituent, var_type = key.split('_')
+            value = data_dict[key][idx]
+            row_dict = {
+                'longitude': longitude,
+                'latitude': latitude,
+                'grid_lon': data_dict['grid_lon'][idx],
+                'grid_lat': data_dict['grid_lat'][idx],
+                'variable': var_type,  # Extracting the variable type (amp, ph, etc.) from the key
+                'constituents': constituent,
+                'type': data_type,  # Extracting the type (u, v, etc.) from the key
+                'value': value  # Store the value
+            }
+            data_list.append(row_dict)
+
+    df = pd.DataFrame(data_list)
+    if 'wide' in mode.lower():
+        return data_to_wide(df, mode)
+
+    return df
+
+
+def get_constituent_vec(
+        dsub, loni, lati, vars=['amp', 'ph'], 
+        constituent=['q1', 'o1', 'p1', 'k1', 'n2', 'm2', 's1', 's2', 'k2', 'm4', 'ms4', 'mn4', '2n2', 'mf', 'mm'], 
+        type=['u', 'v']):
+    out = {'longitude': loni.tolist(), 
+           'latitude': lati.tolist(),
+           'grid_lon': dsub['lon'].values.tolist(),
+           'grid_lat': dsub['lat'].values.tolist()}
+    
+    for TYPE in type:
+        amp_all = dsub[TYPE+'_amp'].values
+        ph_all = dsub[TYPE+'_ph'].values
+        for idx, const in enumerate(constituent):
+            key = f"{TYPE}_{const}"
+            amp = amp_all[..., idx]
+            ph = ph_all[..., idx]
+            cph = -1j * ph * np.pi / 180.0
+            hc = amp * np.exp(cph)
+            if 'amp' in vars:
+                out[key+"_amp"] = amp.tolist()
+            if 'ph' in vars:
+                out[key+"_ph"] = ph.tolist()    
+            if 'hc' in vars:
+                out[key+"_real"] = hc.real.tolist()
+                out[key+"_imag"] = hc.imag.tolist()
+    return out
 
 
 def get_constituent(dz, lon, lat, vars=['amp', 'ph'],
@@ -498,10 +563,10 @@ async def get_tide_const(
             example="25.02,24.82"),
     mode: Optional[str] = Query(
         None,
-        description="Allowed modes: object, row, wide. Optional can be none (default output is in list format), and multiple/special modes can be separated by comma"),
+        description="Allowed modes: object, row, wide. Optional can be none (default output is dataframe in wide format), and multiple/special modes can be separated by comma"),
     tol: Optional[float] = Query(
         None, 
-        description="Tolerance for nearest method to locate points. Nearest method can explictly specied in mode as a special mode, or by just giving tolerance value. Default tolerance is ±1/30 degree, and maximum is ±2.5*1/30 degree."),
+        description="Tolerance for nearest method to locate points. Nearest method can explictly specied in mode as a special mode, or by just giving tolerance value. Default tolerance is ±1/30 degree, and maximum is ±0.25 degree."),
     append: Optional[str] = Query(
         None, description="Data fields to append, separated by commas. If none, 'z': tide height is default. Allowed fields: z, u, v"),
     constituent: Optional[str] = Query(
@@ -557,6 +622,9 @@ async def get_tide_const(
         config.dz.close()
         return JSONResponse(status_code=status.HTTP_400_BAD_REQUEST,
                             content=jsonable_encoder({"Error": "Check your input of lon/lat should be in equal length"}))
+    onlyOnePt = False
+    if len(loni) == 1:
+        onlyOnePt = True   
 
     mlon, mlat = arr_global_lonlat(loni, lati) #to 0-360
 
@@ -578,10 +646,22 @@ async def get_tide_const(
             raise HTTPException(
                 status_code=400, detail="Invalid constituents. Allowed constituents are 'q1','o1','p1','k1','n2','m2','s1','s2','k2','m4','ms4','mn4','2n2','mf','mm'")
 
-    pars = list(set([par.strip() for par in complex.split(
-        ',') if par.strip() in ['amp', 'ph', 'hc']]))
+    if complex is None:
+        complex = 'amp,ph'
+
+    if ',' in complex:       
+        pars = list(set([par.strip() for par in complex.split(
+            ',') if par.strip() in ['amp', 'ph', 'hc']]))
+    elif complex.strip() in ['amp', 'ph', 'hc']:
+        pars=[complex.strip()]
+
     if not pars:
         pars = ['amp', 'ph']
+
+    if mode is None:
+        mode = 'wide'
+    if onlyOnePt:
+        mode = mode + ',onlyOnePt'    
 
     findNear = False
     if 'nearest' in mode.lower():
@@ -591,52 +671,72 @@ async def get_tide_const(
         findNear = True        
         if tol in [np.nan, np.NaN, None] or tol <= 0:
             tol = config.gridSz
-        elif tol > 2.5*config.gridSz:
-            tol = 2.5*config.gridSz
+        elif tol > 7.5*config.gridSz:
+            tol = 7.5*config.gridSz
 
     #pre-subsetting if bounding box within 45 x 45 degrees
-    min_lon, max_lon = min(loni), max(loni)
-    min_lat, max_lat = min(lati), max(lati)
-    lon_rng = max_lon - min_lon
-    lat_rng = max_lat - min_lat
-    if (lon_rng > config.LON_RANGE_LIMIT and lat_rng > config.LAT_RANGE_LIMIT) or lon_rng * lat_rng > config.AREA_LIMIT or np.sign(min_lon) != np.sign(max_lon):
-        # Note if sign is different, do pre-subset may cause error because we must use slice in ds.sel
-        ds = config.dz.sel(constituents=cons)   
-    else:
-        #if np.sign(min_lon) != np.sign(max_lon):
-        #    min_lon, max_lon = min(mlon), max(mlon)   
-        #    subset1 = config.dz.sel(
-        #            lon=slice(min_lon, 360),
-        #            lat=slice(min_lat, max_lat+1.0*config.gridSz),
-        #            constituents=cons)
-        #    subset2 = config.dz.sel(
-        #            lon=slice(0, max_lon+1.0*config.gridSz),
-        #            lat=slice(min_lat, max_lat+1.0*config.gridSz),
-        #            constituents=cons)
-        #    ds = xr.concat([subset1, subset2], dim='lon')
-        #else:    
-        min_lon, max_lon = min(mlon), max(mlon)
-        ds = config.dz.sel(lon=slice(min_lon, max_lon+1.0*config.gridSz),
-                    lat=slice(min_lat, max_lat+1.0*config.gridSz),
-                    constituents=cons)
-       
-    out = []
-    for lon0, lat0 in zip(mlon, mlat):
-        if findNear:
-            dsub = ds.sel(lon=lon0, lat=lat0, method="nearest", tolerance=tol)
+    if not onlyOnePt:
+        min_lon, max_lon = min(loni), max(loni)
+        min_lat, max_lat = min(lati), max(lati)
+        lon_rng = max_lon - min_lon
+        lat_rng = max_lat - min_lat
+        if (lon_rng > config.LON_RANGE_LIMIT and lat_rng > config.LAT_RANGE_LIMIT) or (
+            lon_rng * lat_rng > config.AREA_LIMIT) or np.sign(min_lon) != np.sign(max_lon):
+            # Note if sign is different, do pre-subset may cause error because we must use slice in ds.sel
+            ds = config.dz.sel(constituents=cons)   
         else:
-            dsub = ds.sel(lon=slice(lon0, lon0+1.0*config.gridSz),
-                          lat=slice(lat0, lat0+1.0*config.gridSz))
-        #results = {}
-        #results['longitude'] = lon0
-        #results['latitude'] = lat0
-        #results['grid_lon'] = dsub["lon"].values[0]
-        #results['grid_lat'] = dsub["lat"].values[0]
-        constants = get_constituent(dsub, lon0, lat0, vars=pars, constituent=cons, type=variables)
-        #for key, value in constants.items():
-        #    results[key] = value
-        out.append(constants)
-    #print(out)
+            #if np.sign(min_lon) != np.sign(max_lon):
+            #    min_lon, max_lon = min(mlon), max(mlon)   
+            #    subset1 = config.dz.sel(
+            #            lon=slice(min_lon, 360),
+            #            lat=slice(min_lat, max_lat+1.0*config.gridSz),
+            #            constituents=cons)
+            #    subset2 = config.dz.sel(
+            #            lon=slice(0, max_lon+1.0*config.gridSz),
+            #            lat=slice(min_lat, max_lat+1.0*config.gridSz),
+            #            constituents=cons)
+            #    ds = xr.concat([subset1, subset2], dim='lon')
+            #else:    
+            min_lon, max_lon = min(mlon), max(mlon)
+            ds = config.dz.sel(lon=slice(min_lon-0.5*config.gridSz, max_lon+0.5*config.gridSz),
+                               lat=slice(min_lat-0.5*config.gridSz, max_lat+0.5*config.gridSz),
+                               constituents=cons)        
+        #vectorized version
+        # Create a multi-dimensional coordinate array for vectorized selection
+        coords = xr.DataArray(np.arange(len(mlon)), 
+                    coords={'points_lon': ('points', mlon), 
+                            'points_lat': ('points', mlat)}, dims='points')
+        if findNear:
+            dsub = ds.sel(lon=coords.points_lon, lat=coords.points_lat, method="nearest", tolerance=tol)
+        else:
+            dsub = ds.sel(lon=coords.points_lon, lat=coords.points_lat, method="nearest", tolerance=0.5*config.gridSz)
+
+    else:        
+        if findNear:
+            dsub = config.dz.sel(lon=mlon[0], lat=mlat[0], method="nearest", tolerance=tol)
+        else:
+            dsub = config.dz.sel(lon=slice(mlon[0]-0.5*config.gridSz, mlon[0]+0.5*config.gridSz),
+                                 lat=slice(mlat[0]-0.5*config.gridSz, mlat[0]+0.5*config.gridSz))
+
+    out = get_constituent_vec(dsub, loni, lati, vars=pars, constituent=cons, type=variables)
+    #nested-loop version
+    #out = []
+    #for lon0, lat0 in zip(mlon, mlat):
+    #    if findNear:
+    #        dsub = ds.sel(lon=lon0, lat=lat0, method="nearest", tolerance=tol)
+    #    else:
+    #        dsub = ds.sel(lon=slice(lon0, lon0+1.0*config.gridSz),
+    #                      lat=slice(lat0, lat0+1.0*config.gridSz))
+    #    #results = {}
+    #    #results['longitude'] = lon0
+    #    #results['latitude'] = lat0
+    #    #results['grid_lon'] = dsub["lon"].values[0]
+    #    #results['grid_lat'] = dsub["lat"].values[0]
+    #    constants = get_constituent(dsub, lon0, lat0, vars=pars, constituent=cons, type=variables)
+    #    #for key, value in constants.items():
+    #    #    results[key] = value
+    #    out.append(constants)
+    # print("Test vec version:", out)
 
     if mode is not None and 'object' in mode.lower():
         out_encoded = custom_encoder(out)
@@ -644,7 +744,7 @@ async def get_tide_const(
         # json_data = json.dumps(out_encoded)
         return ORJSONResponse(content=jsonable_encoder(custom_encoder(out)))
 
-    dfout = const_to_output(out, mode)
+    dfout = const_to_output_vec(out, mode)
     #dfout = dfout.where(pd.notna(dfout), None)
     #print(dfout)
     if mode is not None and 'row' in mode.lower():